+++ pkcs12_certificate_extractor/src/main/java/com/sauken/s_fide/pkcs12_certificate_extractor/PKCS12CertificateExtractor.java
@@ -246,6 +246,8 @@ public class PKCS12CertificateExtractor {
     }
 
     private static class CustomException extends RuntimeException {
+        private static final long serialVersionUID = 1L;
+
         public CustomException(String message) {
             super(message);
         }